Refactors how divs are tied to card objects
Card div dom elements are now tied to card objects via a new property
on each card object which holds the ID of the div into which the card
was dealt. This solves the bug where check of board for existing set
before dealing more cards was including hidden divs.